requirement.txt
altgraph==0.17.3
dnspython==2.3.0
numpy==1.24.2
packaging==23.1
pandas==2.0.0
pefile==2023.2.7
Pygments==2.15.1
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
pymongo==4.3.3
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PyQt6==6.5.0
PyQt6-Qt6==6.5.0
PyQt6-sip==13.5.0
PySide6==6.5.0
PySide6-Addons==6.5.0
PySide6-Essentials==6.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32-ctypes==0.2.0
QtAwesome==1.2.3
qtbox==1.2.0
QtPy==2.3.1
shiboken6==6.5.0
six==1.16.0
tzdata==2023.3
